Update to cmake to add library, GUI allows for selecting all files in folder.
