+++ src/encrypter.h
@@ -4,9 +4,6 @@
 #include <iostream>
 #include <string>
 
-#include <iostream>
-#include <string>
-
 using namespace std;
 
 class Encrypter {
